+++ examples/full/src/shared/i18n/locales/en/index.ts
@@ -1,0 +1,5 @@
+import common from './common'
+import crews from './crews'
+import orders from './orders'
+
+export default { common, crews, orders }